Create Dynamic Programming.cpp
T9 - text detector
"T9 is a predictive text technology for mobile phones (specifically those that contain a 3×4 numeric keypad)" - Wikipedia.
Most old phones used this technology to type a message based on clicked numbers. For a given number and the list of words. A program with a dynamic programming approach to returning possible words
that exist in the number.